refactor(issue): move issue comment to separate file
